Conflicts:
	include/thermo_moist.h

+++ include/fields.h
@@ -111,8 +111,8 @@ class cfields
     int checkaddcross(std::string, std::string, std::vector<std::string> *, std::vector<std::string> *);
 
     // masks
-    int calcmaskwplus(double *, double *, int *, int *, double *, double *, double *);
-    int calcmaskwmin (double *, double *, int *, int *, double *, double *, double *);
+    int calcmaskwplus(double *, double *, double *, int *, int *, int *, double *);
+    int calcmaskwmin (double *, double *, double *, int *, int *, int *, double *);
 
     // perturbations
     double rndamp;
