+++ include/thermo_moist.h
@@ -63,15 +63,14 @@ class cthermo_moist : public cthermo
     cstats *stats;
 
     // masks
-    int calcmaskql    (double *, double *, int *, int *, double *, double *, double *);
-    int calcmaskqlcore(double *, double *, int *, int *, double *, double *, double *, double *, double *);
+    int calcmaskql    (double *, double *, double *, int *, int *, int *, double *);
+    int calcmaskqlcore(double *, double *, double *, int *, int *, int *, double *, double *, double *);
 
     int calcbuoyancytend_2nd(double *, double *, double *, double *, double *, double *, double *, double *);
     int calcbuoyancytend_4th(double *, double *, double *, double *, double *, double *, double *, double *);
 
     int calcbuoyancy(double *, double *, double *, double *, double *, double *);
     int calcN2(double *, double *, double *, double *); ///< Calculation of the Brunt-Vaissala frequency.
-
     int calcbasestate(double *, double *, double *, double *, double *, double *, double *, double *, double *, double *);
 
     int calcqlfield(double *, double *, double *, double *);
